+++ cmake/BuildConfiguration.cmake
@@ -53,7 +53,15 @@ if(CMAKE_CXX_COMPILER_ID STREQUAL "Clang"
 endif()
 
 # Set the paths where the executable, libraries and header paths
-set(CMAKE_INSTALL_PREFIX "${CMAKE_BINARY_DIR}/install")
+if ("${CMAKE_VERSION}" VERSION_LESS "3.7.1")
+	# overwrite since the default is system directory
+	set(CMAKE_INSTALL_PREFIX "${CMAKE_BINARY_DIR}/install")
+else ()
+	if (CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
+		# overwrite the default value initialized by CMake
+		set(CMAKE_INSTALL_PREFIX "${CMAKE_BINARY_DIR}/install")
+	endif ()
+endif ()
 set(INSTALL_BIN_DIR "${CMAKE_INSTALL_PREFIX}/bin" CACHE PATH "Installation directory for executables")
 set(INSTALL_LIB_DIR "${CMAKE_INSTALL_PREFIX}/lib" CACHE PATH "Installation directory for libraries")
 set(INSTALL_INC_DIR "${CMAKE_INSTALL_PREFIX}/include" CACHE PATH "Installation directory for headers")