+++ cmake/BuildFunctions.cmake
@@ -99,7 +99,7 @@ endfunction()
 
 ########################################################################################################################
 # Internal utility function to set common flags on any OpenKit target
-function(_set_common_flags target)
+function(_set_common_flags target IS_TEST_TARGET)
 
     # add preprocessor definitions
     if (NOT OPEN_KIT_PREPROCESSOR_DEFINITIONS STREQUAL "")
@@ -112,21 +112,38 @@ function(_set_common_flags target)
         target_compile_definitions(${target} PRIVATE $<$<CONFIG:Release>:${OPEN_KIT_PREPROCESSOR_DEFINITIONS_RELEASE}>)
     endif ()
 
-	_set_compiler_flags(${target} FALSE ${USE_CXX_COMPILER}) ## test_code=FALSE
+	_set_compiler_flags(${target} ${IS_TEST_TARGET} ${USE_CXX_COMPILER})
 
     # add linker options
-    if (NOT OPEN_KIT_LINKER_FLAGS STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS "${FLAGS}")
-    endif ()
-    if (NOT OPEN_KIT_LINKER_FLAGS_DEBUG STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_DEBUG}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS_DEBUG "${FLAGS}")
-    endif ()
-    if (NOT OPEN_KIT_LINKER_FLAGS_RELEASE STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_RELEASE}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS_RELEASE "${FLAGS}")
-    endif ()
+	# note: linker flags and librarian flags have different properties
+	get_target_property(target_type ${target} TYPE)
+	if (target_type STREQUAL "STATIC_LIBRARY")
+		if (NOT OPEN_KIT_LIB_LINKER_FLAGS STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LIB_LINKER_FLAGS}")
+			set_target_properties(${target} PROPERTIES STATIC_LIBRARY_FLAGS "${FLAGS}")
+		endif ()
+		if (NOT OPEN_KIT_LIB_LINKER_FLAGS_DEBUG STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LIB_LINKER_FLAGS_DEBUG}")
+			set_target_properties(${target} PROPERTIES STATIC_LIBRARY_FLAGS_DEBUG "${FLAGS}")
+		endif ()
+		if (NOT OPEN_KIT_LIB_LINKER_FLAGS_RELEASE STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LIB_LINKER_FLAGS_RELEASE}")
+			set_target_properties(${target} PROPERTIES STATIC_LIBRARY_FLAGS_RELEASE "${FLAGS}")
+		endif ()
+	else ()
+		if (NOT OPEN_KIT_LINKER_FLAGS STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS}")
+			set_target_properties(${target} PROPERTIES LINK_FLAGS "${FLAGS}")
+		endif ()
+		if (NOT OPEN_KIT_LINKER_FLAGS_DEBUG STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_DEBUG}")
+			set_target_properties(${target} PROPERTIES LINK_FLAGS_DEBUG "${FLAGS}")
+		endif ()
+		if (NOT OPEN_KIT_LINKER_FLAGS_RELEASE STREQUAL "")
+			string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_RELEASE}")
+			set_target_properties(${target} PROPERTIES LINK_FLAGS_RELEASE "${FLAGS}")
+		endif ()
+	endif ()
 
 	set_target_properties( ${target}
 		PROPERTIES
@@ -135,130 +152,63 @@ function(_set_common_flags target)
 		RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
 	)
 
-endfunction()
-
-########################################################################################################################
-# Internal utility function to set common flags on any OpenKit target
-function(_set_common_flags_tests target)
-
-    # add preprocessor definitions
-    if (NOT OPEN_KIT_PREPROCESSOR_DEFINITIONS STREQUAL "")
-        target_compile_definitions(${target} PRIVATE ${OPEN_KIT_PREPROCESSOR_DEFINITIONS})
-    endif ()
-    if (NOT OPEN_KIT_PREPROCESSOR_DEFINITIONS_DEBUG STREQUAL "")
-        target_compile_definitions(${target} PRIVATE $<$<CONFIG:Debug>:${OPEN_KIT_PREPROCESSOR_DEFINITIONS_DEBUG}>)
-    endif ()
-    if (NOT OPEN_KIT_PREPROCESSOR_DEFINITIONS_RELEASE STREQUAL "")
-        target_compile_definitions(${target} PRIVATE $<$<CONFIG:Release>:${OPEN_KIT_PREPROCESSOR_DEFINITIONS_RELEASE}>)
-    endif ()
-
-	_set_compiler_flags(${target} TRUE ${USE_CXX_COMPILER}) ## test_code=FALSE
-
-    # add linker options
-    if (NOT OPEN_KIT_LINKER_FLAGS STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS "${FLAGS}")
-    endif ()
-    if (NOT OPEN_KIT_LINKER_FLAGS_DEBUG STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_DEBUG}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS_DEBUG "${FLAGS}")
-    endif ()
-    if (NOT OPEN_KIT_LINKER_FLAGS_RELEASE STREQUAL "")
-        string(REPLACE ";" " " FLAGS "${OPEN_KIT_LINKER_FLAGS_RELEASE}")
-        set_target_properties(${target} PROPERTIES LINK_FLAGS_RELEASE "${FLAGS}")
-    endif ()
-
-	set_target_properties( ${target}
-		PROPERTIES
-		ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib"
-		LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
-		RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
-	)
-
-endfunction()
-
-########################################################################################################################
-# Internal utility function for building an OpenKit target
-function(_open_kit_build name includedirs libs)
-
-    target_include_directories(${name} PRIVATE ${includedirs})
-
-    # To support mixing linking in static and dynamic libraries, link each
-    # library in with an extra call to TARGET_LINK_LIBRARIES.
-    foreach (lib "${libs}")
-        target_link_libraries(${name} PRIVATE ${lib})
-    endforeach ()
-
-    # setup some common flags
-    _set_common_flags("${name}")
-
-endfunction()
-
-########################################################################################################################
-# Function used to build a shared library.
-function(open_kit_build_shared_library name includedirs libs)
-
-    message(INFO " Configuring shared library '${name} ('INCLUDEDIRS=${includedirs}; LIBS=${libs}) ")
-    add_library(${name} SHARED ${ARGN})
-    
-    _open_kit_build("${name}" "${includedirs}" "${libs}")
-
-endfunction()
-
-########################################################################################################################
-# Function used to build a static library.
-function(open_kit_build_static_library name includedirs libs)
-
-    message(INFO " Configuring static library '${name}' (INCLUDEDIRS=${includedirs}; LIBS=${libs}) ")
-    add_library(${name} STATIC ${ARGN})
-
-    _open_kit_build("${name}" "${includedirs}" "${libs}")
+    # enforce C++11 standard by default in every of our targets
+    enforce_cxx11_standard(${target})
 
 endfunction()
 
 
 ########################################################################################################################
-# Function to build a library - if OPEN_KIT_SHARED_LIBS is on, it's built as shared library, otherwhise as static lib.
-function(open_kit_build_library name includedirs libs)
+# Function to add a library target - if BUILD_SHARED_LIBS is on, it's added as shared library, otherwhise as static lib.
+# An alias target with Dynatrace:: prefix is created also.
+#
+# This also sets up common compiler flags that are used throughout the whole OpenKit build.
+function(open_kit_add_library name)
     ## check if CFLAGS or CXXFLAGS are required
     _determine_compiler_language(${name} ${ARGN})
 
     if (BUILD_SHARED_LIBS)
-        open_kit_build_shared_library("${name}" "${includedirs}" "${libs}" ${ARGN})
+        message(INFO " Configuring shared library '${name}")
+        add_library(${name} SHARED ${ARGN})
     else ()
-        open_kit_build_static_library("${name}" "${includedirs}" "${libs}" ${ARGN})
+        message(INFO " Configuring static library '${name}'")
+        add_library(${name} STATIC ${ARGN})
     endif ()
 
-    # in previous version of curl this was set on "global" level to all libs except zlib
-    # To avoid breaking name changes, we'll keep the -d suffix on OpenKit libs for debug builds
-    set_target_properties(${name}
-	    PROPERTIES
-        DEBUG_POSTFIX "-d"
+    # add aliased target
+    add_library(
+        Dynatrace::${name}
+        ALIAS ${name}
     )
+
+    # setup some common flags
+    _set_common_flags(${name} FALSE)
 
 endfunction()
 
-
 ########################################################################################################################
-# Function to build an executable
-function(open_kit_build_executable name includedirs libs)
+# Function to add an executable target
+#
+# This also sets up common compiler flags that are used throughout the whole OpenKit build.
+function(open_kit_add_executable name)
 
-    message(INFO " Configuring executable '${name}' (INCLUDEDIRS=${includedirs}; LIBS=${libs}")
+    message(INFO " Configuring executable '${name}'")
 
     ## check if CFLAGS or CXXFLAGS are required
     _determine_compiler_language(${name} ${ARGN})
 
     add_executable(${name} ${ARGN})
 
-    _open_kit_build("${name}" "${includedirs}" "${libs}")
+    # setup some common flags
+    _set_common_flags(${name} FALSE)
 
 endfunction()
 
 ########################################################################################################################
-# Function to build a test application
+# Function to build add a test executable target
 #
 # gtest library and appropriate include directories don't need to be set up, since this is added already.
-function(open_kit_build_test name includedirs libs)
+function(open_kit_add_test name)
 
     if (NOT OPENKIT_BUILD_TESTS)
         message(INFO " Tests are disabled for OpenKit project - skpipping ${name}")
@@ -268,23 +218,14 @@ function(open_kit_build_test name includedirs libs)
     ## check if CFLAGS or CXXFLAGS are required
     _determine_compiler_language(${name} ${ARGN})
 
-    message(INFO " Configuring test '${name}' (INCLUDEDIRS=${includedirs}; LIBS=${libs}")
+    message(INFO " Configuring test '${name}'")
     add_executable(${name} ${ARGN})
 
-	set (test_libs ${libs} gmock_main)
-
-	target_include_directories(${name} SYSTEM PRIVATE ${gtest_SOURCE_DIR}/include)
-	target_include_directories(${name} SYSTEM PRIVATE ${gmock_SOURCE_DIR}/include)
-	target_include_directories(${name} PRIVATE ${includedirs})
-
-    # To support mixing linking in static and dynamic libraries, link each
-    # library in with an extra call to TARGET_LINK_LIBRARIES.
-    foreach (lib "${test_libs}")
-        target_link_libraries(${name} PRIVATE ${lib})
-    endforeach ()
+	# add gmock_main library as testing dependency
+	target_link_libraries(${name} PRIVATE gmock_main)
 
     # setup some common flags
-    _set_common_flags_tests("${name}")
+    _set_common_flags("${name}" TRUE)
 
     # add test for CTest
     add_test(NAME ${name}