+++ cmake/CompilerConfiguration.cmake
@@ -44,6 +44,11 @@ if(MSVC)
     set(OPEN_KIT_LINKER_FLAGS /NOLOGO)
     set(OPEN_KIT_LINKER_FLAGS_DEBUG /DEBUG)
     set(OPEN_KIT_LINKER_FLAGS_RELEASE /LTCG)
+
+	# setup linker flags for static libs
+	set(OPEN_KIT_LIB_LINKER_FLAGS /NOLOGO)
+    set(OPEN_KIT_LIB_LINKER_FLAGS_DEBUG /DEBUG)
+    set(OPEN_KIT_LIB_LINKER_FLAGS_RELEASE /LTCG)
  else()
     # setup preprocessor flags
     set(OPEN_KIT_PREPROCESSOR_DEFINITIONS )
@@ -67,6 +72,11 @@ if(MSVC)
     set(OPEN_KIT_LINKER_FLAGS_DEBUG )
     set(OPEN_KIT_LINKER_FLAGS_RELEASE )
 
+	# setup linker flags for static libs (archives)
+	set(OPEN_KIT_LIB_LINKER_FLAGS )
+    set(OPEN_KIT_LIB_LINKER_FLAGS_DEBUG )
+    set(OPEN_KIT_LIB_LINKER_FLAGS_RELEASE )
+
 	if (OPENKIT_32_BIT)
 		# override default CC and CXX flags
 		foreach (flag_var
@@ -80,15 +90,7 @@ if(MSVC)
 	endif()
 
 	if (OPENKIT_MONOLITHIC_SHARED_LIB)
-		# add -fPIC compiler option, in case we are building a monolithic library
-		foreach (flag_var
-				 CMAKE_CXX_FLAGS CMAKE_CXX_FLAGS_DEBUG CMAKE_CXX_FLAGS_RELEASE
-				 CMAKE_C_FLAGS CMAKE_C_FLAGS_DEBUG CMAKE_C_FLAGS_RELEASE)
-
-			if(NOT (${flag_var} MATCHES "-fPIC"))
-				set(${flag_var} "${${flag_var}} -fPIC")
-			endif()
-		endforeach()
+		set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 	endif()
 endif()